fixed up access

+++ build/usr/local/nagios/bin/deploy_local.sh
@@ -4,7 +4,7 @@ set -o errexit
 
 # ensure this script is run as root
 if [[ $EUID != 0 ]]; then
-  sudo $0
+  sudo "$0"
   exit
 fi 
 
@@ -14,22 +14,25 @@ declare -r NCONF=${NCONF:-"${WWW}/nconf"}
 
 declare -r OUTPUT_DIR="${NCONF}/output"
 declare -r TEMP_DIR="${NCONF}/temp"
-declare -r CONF_ARCHIVE="NagiosConfig.tgz"
-
-declare -r outArchive="${OUTPUT_DIR}/${CONF_ARCHIVE}"
-declare -r tmpArchive="${TEMP_DIR}/${CONF_ARCHIVE}"
-
 
 [[ -e "${TEMP_DIR}" ]] || mkdir -p "${TEMP_DIR}"
 
-# update nagios config if CONF_ARCHIVE is newer than 
-[[ "${outArchive}" -nt "${tmpArchive}" ]] && cp "${outArchive}" "${tmpArchive}"
 
-# re-deploy: 'extract' into /usr/loca/nagois/etc & 'reload' nagios
-echo removing old config
-rm -rf "${NAGIOS_HOME}/etc/Default_collector"
-rm -rf "${NAGIOS_HOME}/etc/global"
-echo adding new config
-tar -xzf "${tmpArchive}" -C "${NAGIOS_HOME}/etc"
-echo restartng 'nagios' service
-/etc/init.d/nagios restart
+declare -r outArchive="${OUTPUT_DIR}/NagiosConfig.tgz"
+declare -r tmpArchive="${TEMP_DIR}/NagiosConfig.tgz"
+
+# update nagios config if CONF_ARCHIVE is newer than 
+if [ -e "$outArchive" ]  && [ "$outArchive" -nt "$tmpArchive" ]; then
+    cp "$outArchive" "$tmpArchive"
+
+    # re-deploy: 'extract' into /usr/loca/nagois/etc & 'reload' nagios
+    echo removing old config
+    rm -rf "${NAGIOS_HOME}/etc/Default_collector"
+    rm -rf "${NAGIOS_HOME}/etc/global"
+    echo adding new config
+    tar -xzf "$tmpArchive" -C "${NAGIOS_HOME}/etc"
+    if [ -e /run/nagios.lock ]; then
+        echo restartng 'nagios' service
+        /etc/init.d/nagios restart
+    fi
+fi
